refactoring(spv-790) addressing comments per review and moving functions into instructed packages
