+++ engine/pike/paike_helpers.go
@@ -1,4 +1,4 @@
-package pike2
+package pike
 
 import (
 	"crypto/hmac"
